Use old-school a.href trick to get URL parts
The `URL` API is not supported in IE11 and this is the next best thing to a regex (although it still involves some regex).

Why this came up:

I was working with an accessibility developer who was testing with JAWS in IE11. They asked “where do I find out about addons” to which I replied “emberobserver.com” to which *they* replied “it doesn’t seem to link to the projects repos…”.