Add a local network cache for benchmarking
This commit adds a simple utility for capturing network requests into local
storage and playing them back later for a subsequent page load. This is useful
for benchmarking where you would like to remove as much variables as possible
(network requests in this case) to keep the runs consistent.

To use this, simply add `?perf.capture` to the URL. Once the page has loaded
you can navigate to `?perf.playback` and see that all network requests were
served from the local cache.

Currently, this requires manually wrapping the places where we are making
network requests (i.e. in the Ember Data adapter) with the special ajax`
function. Ideally, we would want to do this at a lower level by stubbing
`XMLHttpRequest`, which would allow all network requests to be captured
automatically.