Add various performance marker
This adds various userTiming markers and measures that can be viewed
from the "Timeline" tab.

This commit also adds two additional benchmarking modes `?perf.profile`
and `?perf.tracing`. The former wraps the initial render in a CPU
profile. The latter is intended to be used with `chrome-tracing` (and
therefore ember-bench) where it redirects to `about:blank` after the
initial render as the termination signal. Both of these modes implies
`?perf.playback`.