Remove accidental network request from test suite
This test passes, but it does so in a way that's probably not what was intended. The application route makes a real network request, which fails. Then the categories.show route tries to chain off that failed promise and gets the rejection.

This change activates mirage instead, so that a fake network request in the application route succeeds, and then the code within the categories.show route can get exercised and deal with the case where the category doesn't exist.

This error was harmless for users, but it's messing up my embroider testing because when we lazy-load the categories routes, nothing listens to the promise from the application route before it rejects, which causes an unhandled promise error that fails the test.